backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
resend
python-docx
reportlab